requirements.txt
# requirements.txt
# Здесь указываются зависимости проекта.
# Формат: package==version
# Примеры (раскомментируйте нужное или добавьте своё):
# numpy==1.26.4
# pandas==2.2.2
# scikit-learn==1.5.1
# pefile==2023.2.7
# capstone==5.0.1
# yara-python==4.5.1
# rich==13.7.1
# click==8.1.7

# Как обновить файл после установки пакетов:
# 1) Активируйте окружение: .\.venv312\Scripts\Activate.ps1
# 2) Установите пакеты: pip install <имя>
# 3) Зафиксируйте версии: pip freeze > requirements.txt

# Текущие базовые пакеты окружения (можно удалить, если не нужны фиксированные версии):
pip
setuptools
wheel
aiogram>=3.0.0,<4.0.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pandas>=2.2.0
requests>=2.31.0
